code/smartcamZMQ/requirements.txt
numpy
pyzmq
PySide6==6.8.1
pyqtgraph
pims
av
opencv-python-headless

